Rename sign up form components.
