+++ tsconfig.extend.json
@@ -42,7 +42,7 @@
         "./api/*"
       ],
       "aline-signup-form":[
-        "./modules/aline-signup-form"
+        "./modules/SignUpFormComponents"
       ]
     }
   }
